Rewrite using FlameGraphs
