+++ Project.toml
@@ -1,12 +1,13 @@
 name = "ProfileView"
 uuid = "c46f51b8-102a-5cf2-8d2c-8597cb0e0da7"
 author = ["Tim Holy <tim.holy@gmail.com>"]
-version = "0.5.3"
+version = "0.6.0"
 
 [deps]
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+FlameGraphs = "08572546-2f56-4bcf-ba4e-bab62c3a3f89"
 Graphics = "a2bd30eb-e257-5431-a919-1863eab51364"
 Gtk = "4c0ca9eb-093a-5379-98c5-f87ac0bbbf44"
 GtkReactive = "27996c0f-39cd-5cc1-a27a-05f136f946b6"
@@ -16,15 +17,15 @@ Profile = "9abbd945-dff8-562f-b5e8-e1ebf5ef1b79"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [compat]
-Cairo = "0.5, 0.6, 0.7, 0.8, 1"
+Cairo = "0.8, 1"
 Colors = "0.9, 0.10, 0.11"
 FileIO = "1"
-Graphics = "0.2, 0.3, 0.4, 1"
-Gtk = "0.16, 0.17, 0.18, 1"
-GtkReactive = "0.5.1, 0.6, 0.7, 1"
+FlameGraphs = "0.1"
+Graphics = "0.4, 1"
+Gtk = "0.18, 1"
+GtkReactive = "0.7, 1"
 IntervalSets = "0.2, 0.3"
 julia = "0.7, 1"
-
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
