+++ src/ProfileView.jl
@@ -1,32 +1,22 @@
 module ProfileView
 
-using Profile, UUIDs
-using Colors
+using Profile
+using FlameGraphs
+using Base.StackTraces: StackFrame
+using InteractiveUtils
+using Gtk.ShortNames, GtkReactive, Colors, FileIO, IntervalSets
+import Cairo
+using Graphics
 
-import Base: isequal, show
-
-# This allows Revise to correct the location information in profiles
-if VERSION >= v"1.5.0-DEV.9"
-    using Profile: getdict   # ref https://github.com/JuliaLang/julia/pull/34235
-else
-    # Use the definition of getdict from Julia 1.5.0-DEV.9+
-    function getdict(data::Vector{UInt})
-        # Lookup is expensive, so do it only once per ip.
-        udata = unique(data)
-        dict = Profile.LineInfoDict()
-        for ip in udata
-            st = Profile.lookup(convert(Ptr{Cvoid}, ip))
-            # To correct line numbers for moving code, put it in the form expected by
-            # Base.update_stackframes_callback[]
-            stn = map(x->(x, 1), st)
-            try Base.invokelatest(Base.update_stackframes_callback[], stn) catch end
-            dict[UInt64(ip)] = map(first, stn)
-        end
-        return dict
-    end
-end
+using FlameGraphs: Node, NodeData
+using Gtk.GConstants.GdkModifierType: SHIFT, CONTROL, MOD1
 
 export @profview
+
+mutable struct ZoomCanvas
+    bb::BoundingBox  # in user-coordinates
+    c::Canvas
+end
 
 """
     @profview f(args...)
@@ -41,439 +31,174 @@ macro profview(ex)
     end
 end
 
-include("tree.jl")
-include("pvtree.jl")
-
-using .Tree
-using .PVTree
-
-include("svgwriter.jl")
-
-struct TagData
-    ip::UInt
-    status::Int
-end
-const TAGNONE = TagData(UInt(0), -1)
-
-mutable struct ProfileData
-    img
-    lidict
-    imgtags
-    fontsize
-end
-
-const bkg = colorant"white"
-const fontcolor = colorant"black"
-const gccolor = colorant"red"
-const colors = distinguishable_colors(13, [bkg,fontcolor,gccolor],
-                                      lchoices=Float64[65, 70, 75, 80],
-                                      cchoices=Float64[0, 50, 60, 70],
-                                      hchoices=range(0, stop=330, length=24))[4:end]
-
-function have_display()
-    !Sys.isunix() && return true
-    Sys.isapple() && return true
-    return haskey(ENV, "DISPLAY")
-end
-
-function __init__()
-    if (isdefined(Main, :IJulia) && !isdefined(Main, :PROFILEVIEW_USEGTK)) || !have_display()
-        # @eval import ProfileViewSVG
-        include(joinpath(@__DIR__, "ProfileViewSVG.jl"))
-        @eval import .ProfileViewSVG
-        @eval begin
-            view(data = Profile.fetch(); C = false, lidict = nothing, colorgc = true, fontsize = 12, combine = true, pruned = []) = ProfileViewSVG.view(data; C=C, lidict=lidict, colorgc=colorgc, fontsize=fontsize, combine=combine, pruned=pruned)
-        end
-    else
-        # @eval import ProfileViewGtk
-        include(joinpath(@__DIR__, "ProfileViewGtk.jl"))
-        @eval import .ProfileViewGtk
-        @eval begin
-            view(data = Profile.fetch(); C = false, lidict = nothing, colorgc = true, fontsize = 12, combine = true, pruned = []) = ProfileViewGtk.view(data; C=C, lidict=lidict, colorgc=colorgc, fontsize=fontsize, combine=combine, pruned=pruned)
-
-            @doc """
-    closeall()
+"""
+    ProfileView.closeall()
 
 Closes all windows opened by ProfileView.
 """
-            closeall() = ProfileViewGtk.closeall()
-        end
-    end
-end
-
-function prepare(data; C = false, lidict = nothing, colorgc = true, combine = true, pruned = [])
-    bt, uip, counts, lidict, lkup = prepare_data(data, lidict)
-    prepare_image(bt, uip, counts, lidict, lkup, C, colorgc, combine, pruned)
-end
-
-function prepare_data(data, lidict)
-    bt, counts = tree_aggregate(data)
-    if isempty(counts)
-        Profile.warning_empty()
-        error("Nothing to view")
-    end
-    len = Int[length(x) for x in bt]
-    keep = len .> 0
-    if length(data) == Profile.maxlen_data()
-        keep[end] = false
-    end
-    bt = bt[keep]
-    counts = counts[keep]
-    # Display has trouble with very large images. If needed, pretend
-    # we took fewer samples.
-    ncounts = sum(counts)
-    if ncounts > 10^4
-        counts = [floor(Int, c/(ncounts/10^4)) for c in counts]  # uniformly reduce the number of backtraces
-        keep = counts .> 0
-        counts = counts[keep]
-        bt = bt[keep]
-        if isempty(counts)
-            error("No backtraces survived pruning.")
-        end
-    end
-    # Do code address lookups on all unique instruction pointers
-    uip = unique(vcat(bt...))
-    if lidict == nothing
-        lidict = getdict(uip)
-        lkup = map(ip->lidict[ip], uip)
-    else
-        lkup = [lidict[ip] for ip in uip]
-    end
-    bt, uip, counts, lidict, lkup
-end
-
-prepare_data(::Nothing, ::Nothing) = nothing, nothing, nothing, nothing, nothing
-
-function prepare_image(bt, uip, counts, lidict, lkup, C, colorgc, combine,
-                       pruned)
-    nuip = length(uip)
-    isjl = Dict(zip(uip, [all(x->!x.from_c, l) for l in lkup]))
-    isgc = Dict(zip(uip, [any(is_noninferrable_call, l) for l in lkup]))
-    isjl[UInt(0)] = false  # needed for root below
-    isgc[UInt(0)] = false
-    p = Profile.liperm(map(first, lkup))
-    rank = similar(p)
-    rank[p] = 1:length(p)
-    ip2so = Dict(zip(uip, rank))
-    so2ip = Dict(zip(rank, uip))
-    # Build the graph
-    level = 0
-    w = sum(counts)
-    root = Tree.Node(PVData(1:w))
-    PVTree.buildgraph!(root, bt, counts, 0, ip2so, so2ip, lidict)
-    PVTree.setstatus!(root, isgc)
-#     Tree.showedges(STDOUT, root, x -> string(get(lidict, x.ip, "root"), ", hspan = ", x.hspan, ", status = ", x.status))
-#     Tree.showedges(STDOUT, root, x -> string(get(lidict, x.ip, "root"), ", status = ", x.status))
-#     Tree.showedges(STDOUT, root, x -> x.status == 0 ? nothing : string(get(lidict, x.ip, "root"), ", status = ", x.status))
-#     checkidentity(ip2so, so2ip)
-#     checkcontains(root, ip2so, so2ip, lidict)
-#     checkstatus(root, isgc, isjl, C, lidict)
-    counts = zeros(Int, length(uip))
-    if !C
-        pruned_ips = Set()
-        pushpruned!(pruned_ips, pruned, lidict)
-        PVTree.prunegraph!(root, isjl, lidict, ip2so, counts, pruned_ips)
-    end
-#     for ip in uip
-#         println(counts[ip2so[ip]], ": ", lidict[ip])
-#     end
-#     if !C
-#         havegc = any([isgc[ip] for ip in uip])
-#         if havegc
-#             @assert checkprunedgc(root, false)
-#         end
-#     end
-#     println("\nPruned:")
-#     Tree.showedges(STDOUT, root, x -> string(get(lidict, x.ip, "root"), ", status = ", x.status))
-    # Generate a "tagged" image
-    rowtags = Any[fill(TAGNONE, w)]
-    buildtags!(rowtags, root, 1)
-    imgtags = hcat(rowtags...)
-    img = buildimg(imgtags, colors, bkg, gccolor, colorgc, combine, lidict)
-    img, lidict, imgtags
-end
-
-function svgwrite(io::IO, data, lidict; C = false, colorgc = true, fontsize = 12, combine = true, pruned = [])
-    img, lidict, imgtags = prepare(data, C=C, lidict=lidict, colorgc=colorgc, combine=combine, pruned=pruned)
-    pd = ProfileData(img, lidict, imgtags, fontsize)
-    show(io, "image/svg+xml", pd)
-end
-function svgwrite(filename::AbstractString, data, lidict; kwargs...)
-    open(filename, "w") do file
-        svgwrite(file, data, lidict; kwargs...)
+function closeall()
+    for (w, _) in window_wrefs
+        destroy(w)
     end
     nothing
 end
-function svgwrite(io::IO; kwargs...)
-    data, lidict = Profile.retrieve()
-    svgwrite(io, data, lidict; kwargs...)
+
+const window_wrefs = WeakKeyDict{Gtk.GtkWindowLeaf,Nothing}()
+
+"""
+    ProfileView.view([fcolor], data=Profile.fetch(); lidict=nothing, C=false, recur=:off, fontsize=14, kwargs...)
+
+View profiling results. `data` and `lidict` must be a matched pair from `Profile.retrieve()`.
+You have several options to control the output, of which the major ones are:
+
+- `fcolor`: an optional coloration function. The main options are `FlameGraphs.FlameColors`
+  and `FlameGraphs.StackFrameCategory`.
+- `C`: if true, the graph will include stackframes from C code called by Julia.
+- `recur`: on Julia 1.4+, collapse recursive calls (see `Profile.print` for more detail)
+
+See [FlameGraphs](https://github.com/timholy/FlameGraphs.jl) for more information.
+"""
+function view(fcolor, data::Vector{UInt64}=Profile.fetch(); lidict=nothing, C=false, combine=true, recur=:off, pruned=FlameGraphs.defaultpruned, kwargs...)
+    g = flamegraph(data; lidict=lidict, C=C, combine=combine, recur=recur, pruned=pruned)
+    g === nothing && return nothing
+    return view(fcolor, g; kwargs...)
 end
-function svgwrite(filename::AbstractString; kwargs...)
-    data, lidict = Profile.retrieve()
-    svgwrite(filename, data, lidict; kwargs...)
+function view(data::Vector{UInt64}=Profile.fetch(); kwargs...)
+    view(FlameGraphs.default_colors, data; kwargs...)
 end
 
-
-Base.showable(::MIME"image/svg+xml", pd::ProfileData) = true
-
-function show(f::IO, ::MIME"image/svg+xml", pd::ProfileData)
-    img = pd.img
-    lidict = pd.lidict
-    imgtags = pd.imgtags
-    fontsize = pd.fontsize
-
-    ncols, nrows = size(img)
-    leftmargin = rightmargin = 10
-    width = 1000
-    topmargin = 30
-    botmargin = 40
-    rowheight = 15
-    height = ceil(rowheight*nrows + botmargin + topmargin)
-    xstep = (width - (leftmargin + rightmargin)) / ncols
-    ystep = (height - (topmargin + botmargin)) / nrows
-    avgcharwidth = 6  # for Verdana 12 pt font
-    function eschtml(str)
-        s = replace(str, '<' => "&lt;")
-        s = replace(s, '>' => "&gt;")
-        s = replace(s, '&' => "&amp;")
-        s
-    end
-    function printrec(f, samples, xstart, xend, y, tag, rgb)
-        width = xend - xstart
-        li = lidict[tag.ip]
-        info = join(["$(l.func) in $(l.file):$(l.line)" for l in li], "; ")
-        shortinfo = join(["$(l.func) in $(basename(string(l.file))):$(l.line)" for l in li], "; ")
-        info = eschtml(info)
-        shortinfo = eschtml(shortinfo)
-        #if avgcharwidth*3 > width
-        #    shortinfo = ""
-        #elseif length(shortinfo) * avgcharwidth > width
-        #    nchars = int(width/avgcharwidth)-2
-        #    shortinfo = eschtml(info[1:nchars] * "..")
-        #end
-        red = round(Integer,255*rgb.r)
-        green = round(Integer,255*rgb.g)
-        blue = round(Integer,255*rgb.b)
-        print(f, """<rect vector-effect="non-scaling-stroke" x="$xstart" y="$y" width="$width" height="$ystep" fill="rgb($red,$green,$blue)" rx="2" ry="2" data-shortinfo="$shortinfo" data-info="$info"/>\n""")
-        #if shortinfo != ""
-        println(f, """\n<text text-anchor="" x="$(xstart+4)" y="$(y+11.5)" font-size="12" font-family="Verdana" fill="rgb(0,0,0)" ></text>""")
-        # end
+function view(fcolor, g::Node{NodeData}; kwargs...)
+    # Display in a window
+    c = canvas(UserUnit)
+    set_gtk_property!(widget(c), :expand, true)
+    f = Frame(c)
+    tb = Toolbar()
+    bx = Box(:v)
+    push!(bx, tb)
+    push!(bx, f)
+    tb_open = ToolButton("gtk-open")
+    tb_save_as = ToolButton("gtk-save-as")
+    push!(tb, tb_open)
+    push!(tb, tb_save_as)
+    # FIXME: likely have to do `allkwargs` in the two below (add in C, combine, recur)
+    signal_connect(open_cb, tb_open, "clicked", Nothing, (), false, (widget(c),kwargs))
+    signal_connect(save_as_cb, tb_save_as, "clicked", Nothing, (), false, (widget(c),g,kwargs))
+    win = Window("Profile", 800, 600)
+    push!(win, bx)
+    GtkReactive.gc_preserve(win, c)
+    # Register the window with closeall
+    window_wrefs[win] = nothing
+    signal_connect(win, :destroy) do w
+        delete!(window_wrefs, win)
     end
 
-    fig_id = string("fig-", replace(string(uuid4()), "-" => ""))
-    svgheader(f, fig_id, width=width, height=height)
-    # rectangles are on a grid and split across multiple columns (must span similar adjacent ones together)
-    for r in 1:nrows
-        # top of rectangle:
-        y = height - r*ystep - botmargin
-        # local vars:
-        prevtag = TAGNONE
-        xstart = xend = 0.0
-        for c in 1:ncols
-            tag = imgtags[c,r]
-            if prevtag == TAGNONE && prevtag != tag
-                # Very first in span
-                xstart = (c-1) * xstep + leftmargin
-            elseif tag != prevtag && tag != TAGNONE && prevtag != TAGNONE
-                # End of old span and start of new one
-                xend = (c-1) * xstep + leftmargin
-                samples = round(Int, (xend - xstart)/xstep)
-                printrec(f, samples, xstart, xend, y, prevtag, img[c-1,r])
-                xstart = xend
-            elseif tag == TAGNONE && tag != prevtag
-                # at end of span and start of nothing
-                xend = (c-1) * xstep + leftmargin
-                samples = round(Int, (xend - xstart)/xstep)
-                printrec(f, samples, xstart, xend, y, prevtag, img[c-1,r])
-                xstart = 0.0
-            elseif c == ncols && tag != TAGNONE
-                # end of span at last element of row
-                xend = (c-1) * xstep + leftmargin
-                samples = round(Int,(xend - xstart)/xstep)
-                printrec(f, samples, xstart, xend, y, tag, img[c,r])
-                xstart = 0.0
-            else
-                # in middle of span
-            end
-            prevtag = tag
+    viewprof(fcolor, c, g; kwargs...)
+
+    # Ctrl-w and Ctrl-q destroy the window
+    signal_connect(win, "key-press-event") do w, evt
+        if evt.state == CONTROL && (evt.keyval == UInt('q') || evt.keyval == UInt('w'))
+            @async destroy(w)
+            nothing
         end
     end
-    svgfinish(f, fig_id)
+
+    Gtk.showall(win)
 end
 
-function buildtags!(rowtags, parent, level)
-    if isleaf(parent)
-        return
+function viewprof(fcolor, c, g; fontsize=14)
+    img = flamepixels(fcolor, g)
+    tagimg = flametags(g, img)
+    img24 = reverse(RGB24.(img), dims=2)
+    fv = XY(0.0..size(img24,1), 0.0..size(img24,2))
+    zr = Signal(ZoomRegion(fv, fv))
+    sigrb = init_zoom_rubberband(c, zr)
+    sigpd = init_pan_drag(c, zr)
+    sigzs = init_zoom_scroll(c, zr)
+    sigps = init_pan_scroll(c, zr)
+    surf = Cairo.CairoImageSurface(img24)
+    sigredraw = draw(c, zr) do widget, r
+        ctx = getgc(widget)
+        set_coordinates(ctx, r)
+        rectangle(ctx, BoundingBox(r.currentview))
+        set_source(ctx, surf)
+        p = Cairo.get_source(ctx)
+        Cairo.pattern_set_filter(p, Cairo.FILTER_NEAREST)
+        fill(ctx)
     end
-    w = length(rowtags[1])
-    if length(rowtags) < level
-        push!(rowtags, fill(TAGNONE, w))
+    lasttextbb = BoundingBox(1,0,1,0)
+    sigmotion = map(c.mouse.motion) do btn
+        # Repair image from ovewritten text
+        if c.widget.is_realized && c.widget.is_sized
+            ctx = getgc(c)
+            if Graphics.width(lasttextbb) > 0
+                r = value(zr)
+                set_coordinates(ctx, r)
+                rectangle(ctx, lasttextbb)
+                set_source(ctx, surf)
+                p = Cairo.get_source(ctx)
+                Cairo.pattern_set_filter(p, Cairo.FILTER_NEAREST)
+                fill(ctx)
+            end
+            # Write the info
+            xu, yu = btn.position.x, btn.position.y
+            sf = gettag(xu, yu)
+            if sf != StackTraces.UNKNOWN
+                str = string(basename(string(sf.file)), ", ", sf.func, ": line ", sf.line)
+                set_source(ctx, fcolor(:font))
+                Cairo.set_font_face(ctx, "sans-serif $(fontsize)px")
+                xi = value(zr).currentview.x
+                xmin, xmax = minimum(xi), maximum(xi)
+                lasttextbb = deform(Cairo.text(ctx, xu, yu, str, halign = xu < (2xmin+xmax)/3 ? "left" : xu < (xmin+2xmax)/3 ? "center" : "right"), -2, 2, -2, 2)
+            end
+            reveal(c)
+        end
     end
-    t = rowtags[level]
-    for c in parent
-        t[c.data.hspan] .= Ref(TagData(c.data.ip, c.data.status))
-        buildtags!(rowtags, c, level+1)
+    # From a given position, find the underlying tag
+    function gettag(xu, yu)
+        x = ceil(Int, Float64(xu))
+        y = ceil(Int, Float64(yu))
+        Y = size(tagimg, 2)
+        x = max(1, min(x, size(tagimg, 1)))
+        y = max(1, min(y, Y))
+        tagimg[x,Y-y+1]
     end
-end
-
-function buildimg(imgtags, colors, bkg, gccolor, colorgc::Bool, combine::Bool, lidict)
-    w = size(imgtags,1)
-    h = size(imgtags,2)
-    img = fill(bkg, w, h)
-    colorlen = round(Int, length(colors)/2)
-    for j = 1:h
-        coloroffset = colorlen*iseven(j)
-        colorindex = 1
-        lasttag = TAGNONE
-        status = 0
-        first = 0
-        nextcolor = colors[coloroffset + colorindex]
-        for i = 1:w
-            t = imgtags[i,j]
-            if t != TAGNONE
-                if t != lasttag && (lasttag == TAGNONE || !(combine && lidict[lasttag.ip] == lidict[t.ip]))
-                    if first != 0
-                        colorindex = fillrow!(img, j, first:i-1, colorindex, colorlen, nextcolor, gccolor, status & colorgc)
-                        nextcolor = colors[coloroffset + colorindex]
-                        status = t.status
-                    end
-                    first = i
-                    lasttag = t
-                else
-                    status |= t.status
-                end
-            else
-                if first != 0
-                    # We transitioned from tag->none, render the previous range
-                    colorindex = fillrow!(img, j, first:i-1, colorindex, colorlen, nextcolor, gccolor, status & colorgc)
-                    nextcolor = colors[coloroffset + colorindex]
-                    first = 0
-                    lasttag = TAGNONE
+    # Left-click prints the full path, function, and line to the console
+    # Right-click calls the edit() function
+    sigshow = map(c.mouse.buttonpress) do btn
+        if btn.button == 1 || btn.button == 3
+            ctx = getgc(c)
+            xu, yu = btn.position.x, btn.position.y
+            sf = gettag(xu, yu)
+            if sf != StackTraces.UNKNOWN
+                if btn.button == 1
+                    println(sf.file, ", ", sf.func, ": line ", sf.line)
+                elseif btn.button == 3
+                    edit(string(sf.file), sf.line)
                 end
             end
         end
-        if first != 0
-            # We got to the end of a row, render the previous range
-            fillrow!(img, j, first:w, colorindex, colorlen, nextcolor, gccolor, status & colorgc)
-        end
     end
-    img
+    append!(c.preserved, [sigrb, sigpd, sigzs, sigps, sigredraw, sigmotion, sigshow])
+    return nothing
 end
 
-function fillrow!(img, j, rng::UnitRange{Int}, colorindex, colorlen, regcolor, gccolor, status)
-    if status > 0
-        img[rng,j] .= gccolor
-        return colorindex
-    else
-        img[rng,j] .= regcolor
-        return mod1(colorindex+1, colorlen)
-    end
+function open_cb(::Ptr, settings::Tuple)
+    C, kwargs = settings
+    selection = open_dialog("Load profile data", toplevel(c), ("*.jlprof","*"))
+    isempty(selection) && return nothing
+    data, lidict = load(File(format"JLD", selection), "li", "lidict")
+    return view(data; lidict=lidict, C=C, kwargs...)
 end
 
-is_noninferrable_call(f) = f.func == :jl_invoke || f.func == :jl_apply_generic
-
-#### Debugging code
-
-function checkidentity(ip2so, so2ip)
-    for (k,v) in ip2so
-        @assert so2ip[v] == k
-    end
+function save_as_cb(::Ptr, profdata::Tuple)
+    c, data, lidict = profdata
+    selection = save_dialog("Save profile data as JLD file", toplevel(c), ("*.jlprof",))
+    isempty(selection) && return nothing
+    FileIO.save(File(format"JLD", selection), "li", data, "lidict", lidict)
+    nothing
 end
 
-function checkcontains(root, ip2so, so2ip, lidict)
-    flag = contains(root, ip2so)
-    if !all(flag)
-        missing = findall(!flag)
-        println("missing ips:")
-        for i in missing
-            @show i
-            @show so2ip[i]
-            println(lidict[so2ip[i]])
-        end
-        error("Internal error: the tree does not contain all ips")
-    end
-end
-
-# This skips the parent, gets everything else
-# (to avoid a problem with root with ip=0)
-function contains(parent::Node, ip2so::Dict)
-    ret = Array(Bool, 0)
-    contains!(ret, parent, ip2so)
-    @show length(ip2so)
-    @show length(ret)
-    return ret
-end
-
-function contains!(ret, parent::Node, ip2so::Dict)
-    for c in parent
-        indx = ip2so[c.data.ip]
-        setindexsafe!(ret, indx, true)
-        contains!(ret, c, ip2so)
-    end
-end
-
-function setindexsafe!(a, i::Integer, val)
-    if i > length(a)
-        insert!(a, i, val)
-    else
-        a[i] = val
-    end
-end
-
-function checkstatus(parent::Node, isgc::Dict, isjl::Dict, C, lidict)
-    if isgc[parent.data.ip] && parent.data.status == 0
-        @show lidict[parent.data.ip]
-        error("gc should be set, and it isn't")
-    end
-    for c in parent
-        checkstatus(c, isgc, isjl, C, lidict)
-    end
-end
-
-function checkprunedgc(parent::Node, tf::Bool)
-    tf |= parent.data.status > 0
-    if !tf
-        for c in parent
-            tf = checkprunedgc(c, tf)
-        end
-    end
-    tf
-end
-
-function pushpruned!(pruned_ips, pruned, lidict)
-    for (ip, liv) in lidict
-        for li in liv
-            if (li.func, basename(string(li.file))) in pruned
-                push!(pruned_ips, ip)
-                break
-            end
-        end
-    end
-end
-
-## A tree representation
-# Identify and counts repetitions of all unique backtraces
-function tree_aggregate(data::Vector{UInt})
-    iz = findall(iszero, data)  # find the breaks between backtraces
-    treecount = Dict{Vector{UInt},Int}()
-    istart = 1
-    for iend in iz
-        tmp = data[iend - 1 : -1 : istart]
-        treecount[tmp] = get(treecount, tmp, 0) + 1
-        istart = iend + 1
-    end
-    bt = Vector{Vector{UInt}}(undef, 0)
-    counts = Vector{Int}(undef, 0)
-    for (k, v) in treecount
-        if !isempty(k)
-            push!(bt, k)
-            push!(counts, v)
-        end
-    end
-    return (bt, counts)
-end
-
-include("precompile.jl")
-_precompile_()
+# include("precompile.jl")
+# _precompile_()
 
 end
